Add TODO: automatic nil handling on custom parsers
A common issue is that some fields have blank values. Generally,
these should always be returned as nil. It's easy, when defining
a custom parser, to forget to return nil for blank values. Often
this results in NoMethod on nil errors.

This feature would automatically return nil before the parser is
run. Then there would be an option, probably `auto_nil: false` or
something, that the user can set to turn off this behavior if they
know that they want to handle nil values manually.