Refactor parsing implementation
When using the `column` DSL method to define a column, parsers can
be specified two ways: 1) by passing a symbol in the `as` option
which names a built-in parser or a user defined parser, or 2) by
passing a block. Previously, a symbol passed by method 1 would be
wrapped in a Proc so that the logic for parsing values was the same
for both named parsers and parser blocks: just execute the proc.

Now, a symbol passed in this way is no longer wrapped in a proc.
Instead, there is a conditional that checks to see if the parser is
a symbol or a proc. Symbols get passed to `send` to execute the
named parser and procs get instance_exec'ed.

This simplifies the structure of the hash that gets built up as
columns are defined and makes the implementation of a method that
is used for validations: `columns_with_named_parsers`. Also, by
cleaning up the implementation of the `columns` method, it sets
the stage for implementing new options.
the stage