Implement coercion of blank values to nil
Before a value is parsed, it is checked to see if it is blank
(is either already nil or contains only whitespace). If it is
blank then parsing never actually happens and nil is returned as
the value instead. Otherwise, parsing proceeds as normal. The
user can also pass `blanks_as_nil: false` in their column
definition to turn this behavior off.