Change DSL method names
Here's what changed:
- `import` is now `rows`
- `error` is now `errors`

Why: `import` will be required for a future DSL method
that gives the user a chance to define what should happen _around_
the rows being imported (as opposed to what happens for each row).
This is indispensible for things like wrapping an entire import
in a database transaction.

Another small thing: pluralized names were chosen because both
rows and errors can happen multiple times per import and you're
defining what happens for all of them. This is as opposed to
the `column` method, which is used to define a _single_ column.