Adds `import` DSL method
This DSL method allows the user to specify things that should
happen before, after, or around the import. This is useful in
various circumstances:

- Keeping track of when an import has been attempted or completed
- (Eventually) wrapping imports in transactions

Wrapping imports in transactions does not currently work, since
CSVParty currently swallows all errors and sends them to the
`errors` block. This will change with better flow control
mechanisms which are coming soon.