Add mechanism to abort import
This allows the user to abort the entire import. This is useful
when, for example, the import is wrapped in a transaction and
the use wants to rollback that transaction.

Currently, the exception that is raised when the user aborts
the import (AbortedImportError) is silently caught, rather than
bubbled up out of the importer class. An `aborted?` method is
provided to check if an import was aborted, as well as an
`abort_message` method to get the message supplied when aborting.

This implementation may change in the future. I'm not confident
that swallowing the exception is the best route forward. However,
I believe the negative to doing things this way is ameliorated
by the fact that the only way to abort an entire import is by
explicit action on the part of the user, so, in theory, they
are familiar enough with the docs to know that they will need to
use the `aborted?` method to determine if their import was
successful.