Change dependency API
Previously, dependencies were passed as a hash of key/value pairs
as the `dependencies` option when instantiating an importer. This
allowed arbitrary dependencies to be configuerd at runtime.
Generally this approach worked fine, but I think this implementation
is an improvement.

This implementation requires that dependencies be declared in the
class body with `depends_on`. The benefit of doing this is twofold:

1. It makes it more clear to somebody reading the class body that
it requires this dependency to be passed in on instantiation.
Whereas, with the previous API that was opaque. You would have to
notice that one of the parser, rows, errors, or import blocks was
referring to an external dependency to realize that it was required.

2. It makes it possible to raise an error if an importer is instanti-
ated without a required dependency. The previously method would just
raise an error when it got to the part of the import process that
required the dependency and then raise an error.