Use Struct instead of OpenStruct
Struct throws an error when trying to access an attribute that
doesn't exist; OpenStruct does not.

We want this behavior because it will raise an error if a user
tries to access an attribute that doesn't exist, which will make
it easier to debug than if `nil` was silently returned instead
(which would just lead to a NoMethod on Nil error down the line, but
would be at an additional level of remove from the actual bug).