Return nil for invalid Date/Time or format
This commit returns nil if either:

1. The Date or Time to be parsed is invalid, or
2. The strptime format specified for the column is invalid

This was a tough call, which may need to be reconsidered down the
road. I decided on this implementation because it seems more user
friendly to not raise an error when encountering an invalid date.
On the other hand, if you need a date to proceed, but you have an
invalid date, that does seem like a good time to raise an error.