Reserve some column names for internal use
This commit prevents users from naming columns either:

- :unparsed
- :csv_string

These are the names of special attributes on the `row` object
that is passed to the user's row processing block. `row.unparsed`
gives access to the unparsed values from the CSV file and
`row.csv_string` gives access to the raw CSV string.