Improve header finding implementation
When a user either:

A) Doesn't specify a column header
B) Specifies a column header with a regex

We have to find a column header from the CSV file that matches
these regex patterns. Previously, this was happening inside of
the `raise_unless_csv_has_all_columns!` validation method.

However, this step needs to happen whether or not validations
happen, so this commit moves this step into the `import!` method
itself to make this more obvious.