Add row_number to object passed to row processor
This commit adds a row_number attribute to the row object that is
passed to a row processor. So a user can do this:

    rows do |row|
      row.row_number
    end

Technically, with this implementation, a user could just access
the row number without using the row object, since the row number
is stored in an instance variable accessible via the row_number
method. So:

    rows do |row|
      row_number
    end

But I decided to add it to the row object to keep the API uniform.