Use row_number instead of @csv.lineno
This commit sets the stage for intercepting CSV parsing errors in
the future by not depending on the CSV class to report the line
number that is currently being imported. Relying on the CSV class
can be problematic because it doesn't increment it's line number
counter when it encounters a parsing error. However if we are
intercepting these errors, we still need to know what row we are
parsing. This switches to relying on a counter that is kept
separately from the one kept by the CSV class.