Raise error on nonexistent CSV file path
This commit handles the case where a user passes a nonexistent
file path as the CSV data to be imported. It also contains wording
to address the situation where the user tries to import a single
line CSV string.

This can arise because a user can pass a string representing either
a path to a CSV file or CSV data. If the string passed has only a
single line, it is assumed to represent the path to a CSV file. So
if a user tries to parse a single line CSV string, they can
encounte this error.