Fix bug in test
Test expected an error to be raised when trying to import a
malformed row of CSV data. But the malformed data was on the first
data row (second row overall). The CSV class doesn't initialize
headers until the first row of data is shifted from the file.
So the error was being raised while initializing headers, rather
than while importing rows.

The code was working correctly, but this commit fixes the test by
making the second data row the row that contains malformed CSV.