Improve CSV assignment API
Previously, a single initialization option and method were used to
assign the CSV data to be imported, even though there were three
different ways to specify the CSV data:

- A file path
- An IO object
- A String

This commit eliminates this method of assigning CSV data in favor
of separate initialization options and methods for each of the
three forms of CSV data. So, it adds :path, :file, and :content
initialization options and csv_path=, csv_file=, and csv_content=
methods.