Disable assigning CSV after instantiation
This commit removes methods for assigning a csv path, file,
content, or options after an importer has been instantiated. Being
able to do this was not very useful, but removing these methods
will pave the way for further simplification because it allows
code to assume that an instance of an importer always has a valid
CSV object.